+++ src/pages/booking/index.tsx
@@ -1,0 +1,1 @@
+export { default as BookingPage } from './BookingPage';
