+++ src/pages/category/index.tsx
@@ -1,0 +1,1 @@
+export { default as CategoryPage } from './CategoryPage';
