+++ src/pages/comments/index.tsx
@@ -1,0 +1,1 @@
+export { default as CommentsPage } from './CommentsPage';
