+++ src/pages/feed/index.tsx
@@ -1,0 +1,1 @@
+export { default as FeedPage } from './FeedPage';
