+++ src/pages/index.tsx
@@ -1,24 +1,23 @@
-export { default as HomePage } from './HomePage';
-export { default as ProductsDetailPage } from './ProductsDetailPage';
+export { HomePage } from './home';
+export { ProductsDetailPage } from './products';
 export { default as WorkshopPage } from './WorkshopPage/WorkshopPage';
-export { default as CategoryPage } from './CategoryPage';
-export { default as FeedPage } from './FeedPage';
+export { CategoryPage } from './category';
+export { FeedPage } from './feed';
 export { default as LoginPage } from './LoginPage';
 export { default as RedirectPageAfterKakao } from './LoginPage/RedirectPageAfterKakao';
 export { default as SignupCheckLocation } from './LoginPage/SignupCheckLocation';
 export { default as SignupAuthorInfo } from './LoginPage/SignupAuthorInfo';
-export { default as BookingPage } from './BookingPage';
-export { default as SearchPage } from './SearchPage';
+export { BookingPage } from './booking';
+
+export { SearchPage, SearchResultsPage } from './search';
 export {
   UserPage,
   UserBookedPage,
   UserFollowingPage,
   UserProfileEditPage,
   UserReviewPage,
-} from './UserPage';
-export { default as NotFoundPage } from './NotFoundPage';
-export { default as CommentsPage } from './CommentsPage';
-export { default as SearchResultsPage } from './SearchResultsPage';
-export { default as UploadFeedPage } from './UploadFeedPage';
-export { default as UploadCommentPage } from './UploadCommentPage';
-export { default as UploadProductPage } from './UploadProductPage';
+} from './user';
+export { NotFoundPage } from './notFound';
+export { CommentsPage } from './comments';
+
+export { UploadFeedPage, UploadCommentPage, UploadProductPage } from './upload';
