+++ src/pages/notFound/index.tsx
@@ -1,0 +1,1 @@
+export { default as NotFoundPage } from './NotFoundPage';
