+++ src/pages/products/index.tsx
@@ -1,0 +1,1 @@
+export { default as ProductsDetailPage } from './ProductsDetailPage';
