+++ src/pages/search/index.tsx
@@ -1,0 +1,2 @@
+export { default as SearchPage } from './SearchPage';
+export { default as SearchResultsPage } from './SearchResultsPage';
