+++ src/pages/upload/index.tsx
@@ -1,0 +1,3 @@
+export { default as UploadCommentPage } from './UploadCommentPage';
+export { default as UploadFeedPage } from './UploadFeedPage';
+export { default as UploadProductPage } from './UploadProductPage';
